+++ LTX_2_MLX/pipelines/distilled.py
@@ -63,7 +63,7 @@ class DistilledConfig:
     tiling_config: Optional[TilingConfig] = None
 
     # Compute settings
-    dtype: mx.Dtype = mx.float32
+    dtype: mx.Dtype = mx.bfloat16
 
     # Audio configuration
     audio_enabled: bool = False
